+++ src/features/auth/auth.routes.ts
@@ -1,0 +1,6 @@
+import { Router } from 'express';
+
+import { grantAuth } from './auth.controllers.js';
+
+export const authRoutes = Router();
+authRoutes.post('/grant', grantAuth);
