+++ src/index.ts
@@ -1,10 +1,9 @@
 import cors from 'cors';
 import express from 'express';
 
+import { authRoutes } from './features/auth/auth.routes.js';
 import { listImages } from './features/images/images.controllers.js';
 import { imagesRoutes } from './features/images/images.routes.js';
-import { grantAuth } from './shared/auth.js';
-// import { requireAuth } from './shared/auth.js';
 import { config } from './shared/config.js';
 
 const app = express();
@@ -17,15 +16,13 @@ app.use(
   }),
 );
 app.use(express.json());
-// app.use(requireAuth);
-
-app.post('/api/auth/grant', grantAuth);
 
 /**
  * @deprecated use api/images/list instead
  */
 app.get('/api/listImages', listImages);
 app.use('/api/images', imagesRoutes);
+app.use('/api/auth', authRoutes);
 
 app.get('/api/hello', (req, res) => {
   res.json({ message: 'Hello from backend!' });
